Converted part of the Processing code
This is a port of Processing to C++ of all the code that is not
dependant on the Processing API.  Specifically, drawing code is not
included, but the functional bits are in place.  It is untested at this
time, but it mostly converted line-for-line just fine so I don't expect
problems.