Ghost ships implemented
It has a definite sprite position, but it draws 4 ghosts
up/down/left/right of it so the transition on the edges is smoother.  It
doesn't show up on the opposite corner when it is in a corner, but this
is kind of an edge case that won't appear much once the game gets going.